alpaca-trading-bot/requirements.txt
alpaca-trade-api==3.0.2
websocket-client==1.6.4
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Desktop GUI dependencies
PyQt6==6.6.1
PyQt6-Charts==6.6.0
pyqtgraph==0.13.3
qasync==0.27.1
